+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Configuration:
 
+ * `amb-sidecar`: The default value of `USE_STATSD` has changed from `true` to `false`.
  * Bump license key schema v0 → v1.  The developer portal requires a
    v1 license with the "devportal" feature enabled.  Some future
    version of the other functionality will drop support for v0 license
@@ -17,6 +18,7 @@ Configuration:
  * Everything is now on one port: `APRO_HTTP_PORT`, which defaults to `8500`.
  * `LOG_LEVEL` no longer exists; everything obeys `APP_LOG_LEVEL`.
  * The meaning of `REDIS_POOL_SIZE` has changed slightly; there are no longer separate connection pools for ratelimit and filtering; the maxiumum number of connections is now `REDIS_POOL_SIZE` instead of 2×`REDIS_POOL_SIZE`.
+ * The `amb-sidecar` RateLimitService can now report to statsd, and attempts to do so by default (`USE_STATSD`, `STATSD_HOST`, `STATSD_PORT`, `GOSTATS_FLUSH_INTERVAL_SECONDS`).
 
 Behavior: